# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Locked — Mental performance app for athletes
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  clerkId: string;
  sport: string;
  displayName: string;
  createdAt: string;
}

export interface UserInput {
  /** @minLength 1 */
  sport: string;
  /** @minLength 1 */
  displayName: string;
}

export interface Checkin {
  id: number;
  clerkId: string;
  sport: string;
  /**
     * @minimum 1
     * @maximum 10
     */
  focusScore: number;
  /**
     * @minimum 1
     * @maximum 10
     */
  confidenceScore: number;
  /**
     * @minimum 1
     * @maximum 10
     */
  energyScore: number;
  /** @nullable */
  note: string | null;
  createdAt: string;
}

export interface CheckinInput {
  /**
     * @minimum 1
     * @maximum 10
     */
  focusScore: number;
  /**
     * @minimum 1
     * @maximum 10
     */
  confidenceScore: number;
  /**
     * @minimum 1
     * @maximum 10
     */
  energyScore: number;
  note?: string;
}

export interface CheckinList {
  checkins: Checkin[];
  total: number;
  page: number;
  limit: number;
}

export interface TodayCheckin {
  checkin: Checkin | null;
}

export interface FeedItem {
  id: number;
  sport: string;
  focusScore: number;
  confidenceScore: number;
  energyScore: number;
  /** @nullable */
  note: string | null;
  createdAt: string;
}

export interface FeedList {
  items: FeedItem[];
  total: number;
  page: number;
  limit: number;
}

export interface PersonalStats {
  totalCheckins: number;
  currentStreak: number;
  longestStreak: number;
  avgFocus: number;
  avgConfidence: number;
  avgEnergy: number;
  avgOverall: number;
}

export interface TrendPoint {
  date: string;
  focusScore: number;
  confidenceScore: number;
  energyScore: number;
  overall: number;
}

export type ListMyCheckinsParams = {
page?: number;
limit?: number;
};

export type GetCommunityFeedParams = {
sport?: string;
page?: number;
limit?: number;
};