# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Locked — Mental performance app for athletes
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: users
    description: User profile operations
  - name: checkins
    description: Daily check-in operations
  - name: feed
    description: Community feed operations
  - name: stats
    description: Personal stats and patterns
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
    put:
      operationId: upsertMe
      tags: [users]
      summary: Create or update current user profile (set sport, display name)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Validation error
        "401":
          description: Unauthorized

  /checkins:
    get:
      operationId: listMyCheckins
      tags: [checkins]
      summary: Get current user's check-in history (paginated)
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: List of check-ins
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckinList"
        "401":
          description: Unauthorized
    post:
      operationId: createCheckin
      tags: [checkins]
      summary: Submit today's mental check-in (once per day)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckinInput"
      responses:
        "201":
          description: Created check-in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Checkin"
        "400":
          description: Already checked in today or validation error
        "401":
          description: Unauthorized

  /checkins/today:
    get:
      operationId: getTodayCheckin
      tags: [checkins]
      summary: Get today's check-in (null if not done yet)
      responses:
        "200":
          description: Today's check-in or null
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodayCheckin"
        "401":
          description: Unauthorized

  /feed:
    get:
      operationId: getCommunityFeed
      tags: [feed]
      summary: Get anonymous community check-ins for a sport
      parameters:
        - name: sport
          in: query
          required: false
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Community feed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedList"

  /stats/me:
    get:
      operationId: getMyStats
      tags: [stats]
      summary: Get personal stats — streak, averages, weekly trend
      responses:
        "200":
          description: Personal stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonalStats"
        "401":
          description: Unauthorized

  /stats/me/trend:
    get:
      operationId: getMyTrend
      tags: [stats]
      summary: Get last 30 days of check-in score trend (for charts)
      responses:
        "200":
          description: Trend data points
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrendPoint"
        "401":
          description: Unauthorized

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    User:
      type: object
      required: [clerkId, sport, displayName, createdAt]
      properties:
        clerkId:
          type: string
        sport:
          type: string
        displayName:
          type: string
        createdAt:
          type: string
          format: date-time

    UserInput:
      type: object
      required: [sport, displayName]
      properties:
        sport:
          type: string
          minLength: 1
        displayName:
          type: string
          minLength: 1

    Checkin:
      type: object
      required: [id, clerkId, sport, focusScore, confidenceScore, energyScore, note, createdAt]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        sport:
          type: string
        focusScore:
          type: integer
          minimum: 1
          maximum: 10
        confidenceScore:
          type: integer
          minimum: 1
          maximum: 10
        energyScore:
          type: integer
          minimum: 1
          maximum: 10
        note:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    CheckinInput:
      type: object
      required: [focusScore, confidenceScore, energyScore]
      properties:
        focusScore:
          type: integer
          minimum: 1
          maximum: 10
        confidenceScore:
          type: integer
          minimum: 1
          maximum: 10
        energyScore:
          type: integer
          minimum: 1
          maximum: 10
        note:
          type: string

    CheckinList:
      type: object
      required: [checkins, total, page, limit]
      properties:
        checkins:
          type: array
          items:
            $ref: "#/components/schemas/Checkin"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    TodayCheckin:
      type: object
      required: [checkin]
      properties:
        checkin:
          oneOf:
            - $ref: "#/components/schemas/Checkin"
            - type: "null"

    FeedItem:
      type: object
      required: [id, sport, focusScore, confidenceScore, energyScore, note, createdAt]
      properties:
        id:
          type: integer
        sport:
          type: string
        focusScore:
          type: integer
        confidenceScore:
          type: integer
        energyScore:
          type: integer
        note:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    FeedList:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/FeedItem"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    PersonalStats:
      type: object
      required: [totalCheckins, currentStreak, longestStreak, avgFocus, avgConfidence, avgEnergy, avgOverall]
      properties:
        totalCheckins:
          type: integer
        currentStreak:
          type: integer
        longestStreak:
          type: integer
        avgFocus:
          type: number
        avgConfidence:
          type: number
        avgEnergy:
          type: number
        avgOverall:
          type: number

    TrendPoint:
      type: object
      required: [date, focusScore, confidenceScore, energyScore, overall]
      properties:
        date:
          type: string
          format: date
        focusScore:
          type: number
        confidenceScore:
          type: number
        energyScore:
          type: number
        overall:
          type: number