# lib/api-zod/src/generated/types/checkinInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Locked — Mental performance app for athletes
 * OpenAPI spec version: 0.1.0
 */

export interface CheckinInput {
  /**
     * @minimum 1
     * @maximum 10
     */
  focusScore: number;
  /**
     * @minimum 1
     * @maximum 10
     */
  confidenceScore: number;
  /**
     * @minimum 1
     * @maximum 10
     */
  energyScore: number;
  note?: string;
}