# lib/api-zod/src/generated/types/feedItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Locked — Mental performance app for athletes
 * OpenAPI spec version: 0.1.0
 */

export interface FeedItem {
  id: number;
  sport: string;
  focusScore: number;
  confidenceScore: number;
  energyScore: number;
  /** @nullable */
  note: string | null;
  createdAt: Date;
}