# lib/api-zod/src/generated/types/feedList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Locked — Mental performance app for athletes
 * OpenAPI spec version: 0.1.0
 */
import type { FeedItem } from './feedItem';

export interface FeedList {
  items: FeedItem[];
  total: number;
  page: number;
  limit: number;
}