# lib/api-zod/src/generated/types/personalStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Locked — Mental performance app for athletes
 * OpenAPI spec version: 0.1.0
 */

export interface PersonalStats {
  totalCheckins: number;
  currentStreak: number;
  longestStreak: number;
  avgFocus: number;
  avgConfidence: number;
  avgEnergy: number;
  avgOverall: number;
}