# lib/api-zod/src/generated/types/todayCheckin.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Locked — Mental performance app for athletes
 * OpenAPI spec version: 0.1.0
 */
import type { Checkin } from './checkin';

export interface TodayCheckin {
  checkin: Checkin | null;
}