# lib/api-zod/src/generated/types/trendPoint.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Locked — Mental performance app for athletes
 * OpenAPI spec version: 0.1.0
 */

export interface TrendPoint {
  date: Date;
  focusScore: number;
  confidenceScore: number;
  energyScore: number;
  overall: number;
}