# lib/api-zod/src/generated/types/userInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Locked — Mental performance app for athletes
 * OpenAPI spec version: 0.1.0
 */

export interface UserInput {
  /** @minLength 1 */
  sport: string;
  /** @minLength 1 */
  displayName: string;
}